Added authentication logic to custom data group

+++ segosarem-backend/src/main/java/com/segosarem/web/webservices/api/CustomDataGroupAPI.java
@@ -16,7 +16,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 //Services
 import com.segosarem.web.webservices.db.service.CustomDataGroupService;
-
+import com.segosarem.web.webservices.db.service.AuthenticationService;
+import com.segosarem.web.webservices.db.service.CommonServiceUtils;
 //Beans
 import com.segosarem.web.webservices.bean.GeneralWsResponseBean;
 
@@ -40,6 +41,9 @@ public class CustomDataGroupAPI {
 	// Services
 	@Autowired
 	CustomDataGroupService customDataGroupService;
+
+	@Autowired
+	AuthenticationService authenticationService;
 	// Services
 
 	// Outlets
@@ -51,8 +55,17 @@ public class CustomDataGroupAPI {
 	@ResponseBody
 	@RequestMapping(method = RequestMethod.POST, value = "/getcustomDataGroupList")
 	public GeneralWsResponseBean getCustomDataGroupList(HttpServletRequest request, HttpServletResponse response) {
-
-		return (GeneralWsResponseBean) customDataGroupService.getAllCustomDataGroup();
+		
+		String token = request.getHeader("token");
+        String latestToken = authenticationService.getLatestToken();
+		Boolean allowContinue = CommonAPIUtils.checkToken(latestToken, token);
+		
+		if(allowContinue) {
+			return (GeneralWsResponseBean) customDataGroupService.getAllCustomDataGroup();
+		}
+		else {
+			return CommonServiceUtils.generateResponseBeanWithUnauthorizedStatus();
+		}
 	}
 
 	// get csgroup by page setting id
@@ -65,23 +78,33 @@ public class CustomDataGroupAPI {
 	@RequestMapping(method = RequestMethod.POST, value = "/getCdGroupByPageStgId", consumes = { "application/json" })
 	public GeneralWsResponseBean getCdGroupByPageStgId(HttpServletRequest request, HttpServletResponse response,
 			@RequestBody CustomDataGroupBean requestBean) {
-		return (GeneralWsResponseBean) customDataGroupService.getCdGroupByPageStgId(requestBean.getPageSettingId());
+		
+		String token = request.getHeader("token");
+		String latestToken = authenticationService.getLatestToken();
+		Boolean allowContinue = CommonAPIUtils.checkToken(latestToken, token);
+		
+		if(allowContinue) {
+			return (GeneralWsResponseBean) customDataGroupService.getCdGroupByPageStgId(requestBean.getPageSettingId());
+		}
+		else {
+			return CommonServiceUtils.generateResponseBeanWithUnauthorizedStatus();
+		}
 	}
 	// get csgroup by page setting id
 
-	@ApiOperation(value = "Get custom data group by id", response = GeneralWsResponseBean.class)
-	@ApiResponses(value = { @ApiResponse(code = 200, message = "Successfully Get the details"),
-			@ApiResponse(code = 401, message = "You are not authorized to view the resource"),
-			@ApiResponse(code = 403, message = "Accessing the resource you were trying to reach is forbidden"),
-			@ApiResponse(code = 404, message = "The resource you were trying to reach is not found") })
-	@ResponseBody
-	@RequestMapping(method = RequestMethod.POST, value = "/getCustomDataGroupById", consumes = {
-			"application/json" }, produces = { "application/json" })
-	public GeneralWsResponseBean getCustomDataGroupById(HttpServletRequest request, HttpServletResponse response,
-			@RequestBody CustomDataGroupBean requestBean) {
+	// @ApiOperation(value = "Get custom data group by id", response = GeneralWsResponseBean.class)
+	// @ApiResponses(value = { @ApiResponse(code = 200, message = "Successfully Get the details"),
+	// 		@ApiResponse(code = 401, message = "You are not authorized to view the resource"),
+	// 		@ApiResponse(code = 403, message = "Accessing the resource you were trying to reach is forbidden"),
+	// 		@ApiResponse(code = 404, message = "The resource you were trying to reach is not found") })
+	// @ResponseBody
+	// @RequestMapping(method = RequestMethod.POST, value = "/getCustomDataGroupById", consumes = {
+	// 		"application/json" }, produces = { "application/json" })
+	// public GeneralWsResponseBean getCustomDataGroupById(HttpServletRequest request, HttpServletResponse response,
+	// 		@RequestBody CustomDataGroupBean requestBean) {
 
-		return (GeneralWsResponseBean) customDataGroupService.getCustomDataGroupById(requestBean.getCdGroupId());
-	}
+	// 	return (GeneralWsResponseBean) customDataGroupService.getCustomDataGroupById(requestBean.getCdGroupId());
+	// }
 
 	// Supposed to be for secure access only
 	@ApiOperation(value = "Add custom data group", response = GeneralWsResponseBean.class)
@@ -95,7 +118,16 @@ public class CustomDataGroupAPI {
 	public GeneralWsResponseBean addCustomDataGroup(HttpServletRequest request, HttpServletResponse response,
 			@RequestBody CustomDataGroupBean requestBean) {
 
-		return (GeneralWsResponseBean) customDataGroupService.addCustomDataGroup(requestBean);
+		String token = request.getHeader("token");
+		String latestToken = authenticationService.getLatestToken();
+		Boolean allowContinue = CommonAPIUtils.checkToken(latestToken, token);
+		
+		if(allowContinue) {
+			return (GeneralWsResponseBean) customDataGroupService.addCustomDataGroup(requestBean);
+		}
+		else {
+			return CommonServiceUtils.generateResponseBeanWithUnauthorizedStatus();
+		}
 	}
 
 	@ApiOperation(value = "Update custom data group / Also can be used to update status to either active, deactive, or deleted", response = GeneralWsResponseBean.class)
@@ -109,7 +141,16 @@ public class CustomDataGroupAPI {
 	public GeneralWsResponseBean updateCustomDataGroup(HttpServletRequest request, HttpServletResponse response,
 			@RequestBody CustomDataGroupBean requestBean) {
 
-		return (GeneralWsResponseBean) customDataGroupService.updateCustomDataGroup(requestBean);
+		String token = request.getHeader("token");
+		String latestToken = authenticationService.getLatestToken();
+		Boolean allowContinue = CommonAPIUtils.checkToken(latestToken, token);
+		
+		if(allowContinue) {
+			return (GeneralWsResponseBean) customDataGroupService.updateCustomDataGroup(requestBean);
+		}
+		else {
+			return CommonServiceUtils.generateResponseBeanWithUnauthorizedStatus();
+		}
 	}
 
 	@ApiOperation(value = "Delete custom data group / Effectively removed it from db", response = GeneralWsResponseBean.class)
@@ -123,6 +164,15 @@ public class CustomDataGroupAPI {
 	public GeneralWsResponseBean deleteCustomDataGroup(HttpServletRequest request, HttpServletResponse response,
 			@RequestBody DeleteEntityReqBean requestBean) {
 
-		return (GeneralWsResponseBean) customDataGroupService.deleteCustomDataGroup(requestBean);
+		String token = request.getHeader("token");
+		String latestToken = authenticationService.getLatestToken();
+		Boolean allowContinue = CommonAPIUtils.checkToken(latestToken, token);
+		
+		if(allowContinue) {
+			return (GeneralWsResponseBean) customDataGroupService.deleteCustomDataGroup(requestBean);
+		}
+		else {
+			return CommonServiceUtils.generateResponseBeanWithUnauthorizedStatus();
+		}
 	}
 }
